[nextest-runner] move FormatVersionError to errors.rs
All errors go in this central location.
